[PvP] Moving user options OUT of PvE
