+++ WrathCombo/Combos/PvE/BLM/BLM_Config.cs
@@ -4,7 +4,6 @@ using ECommons.GameHelpers;
 using ECommons.ImGuiMethods;
 using ImGuiNET;
 using System;
-using WrathCombo.Combos.PvP;
 using WrathCombo.CustomComboNS.Functions;
 using WrathCombo.Extensions;
 using static WrathCombo.CustomComboNS.Functions.CustomComboFunctions;
@@ -124,61 +123,6 @@ internal partial class BLM
                 case CustomComboPreset.BLM_AoE_Thunder:
                     DrawSliderInt(0, 10, BLM_AoE_ThunderHP,
                         $"Stop Using {Thunder2.ActionName()} When Target HP% is at or Below (Set to 0 to Disable This Check)");
-
-                    break;
-
-                // PvP
-
-                // Movement Threshold
-                case CustomComboPreset.BLMPvP_BurstMode:
-                    DrawHorizontalRadioButton(BLMPvP.Config.BLMPVP_BurstButtonOption, "One Button Mode", "Combines Fire & Blizzard onto one button", 0);
-                    DrawHorizontalRadioButton(BLMPvP.Config.BLMPVP_BurstButtonOption, "Dual Button Mode", "Puts the combo onto separate Fire & Blizzard buttons, which will only use that element.", 1);
-
-                    if (BLMPvP.Config.BLMPVP_BurstButtonOption == 0)
-                    {
-                        ImGui.Indent();
-                        DrawRoundedSliderFloat(0.1f, 3, BLMPvP.Config.BLMPvP_Movement_Threshold, "Movement Threshold", 137);
-                        ImGui.Unindent();
-                        if (ImGui.IsItemHovered())
-                        {
-                            ImGui.BeginTooltip();
-                            ImGui.TextUnformatted("When under the effect of Astral Fire, must be\nmoving this long before using Blizzard spells.");
-                            ImGui.EndTooltip();
-                        }
-                    }
-                    break;
-
-                // Burst
-                case CustomComboPreset.BLMPvP_Burst:
-                    DrawAdditionalBoolChoice(BLMPvP.Config.BLMPvP_Burst_SubOption, "Defensive Burst",
-                        "Also uses Burst when under 50%% HP.\n- Will not use outside combat.");
-
-                    break;
-
-                // Elemental Weave
-                case CustomComboPreset.BLMPvP_ElementalWeave:
-                    DrawSliderInt(10, 100, BLMPvP.Config.BLMPvP_ElementalWeave_PlayerHP, "Player HP%", 180);
-                    ImGui.Spacing();
-                    DrawAdditionalBoolChoice(BLMPvP.Config.BLMPvP_ElementalWeave_SubOption, "Defensive Elemental Weave",
-                        "When under, uses Wreath of Ice instead.\n- Will not use outside combat.");
-
-                    break;
-
-                // Lethargy
-                case CustomComboPreset.BLMPvP_Lethargy:
-                    DrawSliderInt(10, 100, BLMPvP.Config.BLMPvP_Lethargy_TargetHP, "Target HP%", 180);
-                    ImGui.Spacing();
-                    DrawAdditionalBoolChoice(BLMPvP.Config.BLMPvP_Lethargy_SubOption, "Defensive Lethargy",
-                        "Also uses Lethargy when under 50%% HP.\n- Uses only when targeted by enemy.");
-
-                    break;
-
-                // Xenoglossy
-                case CustomComboPreset.BLMPvP_Xenoglossy:
-                    DrawSliderInt(10, 100, BLMPvP.Config.BLMPvP_Xenoglossy_TargetHP, "Target HP%", 180);
-                    ImGui.Spacing();
-                    DrawAdditionalBoolChoice(BLMPvP.Config.BLMPvP_Xenoglossy_SubOption, "Defensive Xenoglossy",
-                        "Also uses Xenoglossy when under 50%% HP.");
 
                     break;
             }
