+++ WrathCombo/Combos/PvE/DRK/DRK_Config.cs
@@ -3,7 +3,6 @@
 using ECommons.ImGuiMethods;
 using ImGuiNET;
 using System.Numerics;
-using WrathCombo.Combos.PvP;
 using WrathCombo.CustomComboNS.Functions;
 using WrathCombo.Data;
 using WrathCombo.Window.Functions;
@@ -313,18 +312,6 @@ internal partial class DRK
                         itemWidth: biggest, sliderIncrement: SliderIncrements.Fives);
 
                     break;
-
-                #region PVP
-
-                case CustomComboPreset.DRKPvP_Shadowbringer:
-                    UserConfig.DrawSliderInt(20, 100,
-                        DRKPvP.Config.ShadowbringerThreshold,
-                        "HP% to be at or Above to use ",
-                        itemWidth: 150f, sliderIncrement: SliderIncrements.Fives);
-
-                    break;
-
-                #endregion
 
                 #region One-Button Mitigation
