+++ WrathCombo/Combos/PvE/PCT/PCT_Config.cs
@@ -1,5 +1,4 @@
 using ImGuiNET;
-using WrathCombo.Combos.PvP;
 using WrathCombo.CustomComboNS.Functions;
 using WrathCombo.Extensions;
 using WrathCombo.Window.Functions;
@@ -135,17 +134,6 @@ internal partial class PCT
 
                 case CustomComboPreset.PCT_Variant_Cure:
                     UserConfig.DrawSliderInt(1, 100, PCT_VariantCure, "HP% to be at or under", 200);
-
-                    break;
-
-                // PvP
-                case CustomComboPreset.PCTPvP_BurstControl:
-                    UserConfig.DrawSliderInt(1, 100, PCTPvP.Config.PCTPvP_BurstHP, "Target HP%", 200);
-
-                    break;
-
-                case CustomComboPreset.PCTPvP_TemperaCoat:
-                    UserConfig.DrawSliderInt(1, 100, PCTPvP.Config.PCTPvP_TemperaHP, "Player HP%", 200);
 
                     break;
             }
