+++ WrathCombo/Combos/PvP/ASTPVP.cs
@@ -1,6 +1,9 @@
-﻿using WrathCombo.Core;
+﻿using ECommons.GameHelpers;
+using WrathCombo.Core;
 using WrathCombo.CustomComboNS;
-using static WrathCombo.Combos.PvE.AST;
+using WrathCombo.CustomComboNS.Functions;
+using static WrathCombo.Window.Functions.SliderIncrements;
+using static WrathCombo.Window.Functions.UserConfig;
 
 namespace WrathCombo.Combos.PvP
 {
@@ -32,6 +35,29 @@ namespace WrathCombo.Combos.PvP
                     RetrogradeReady = 4331;
 
         }
+
+        public class Config
+        {
+            public static UserInt
+                ASTPvP_Burst_PlayCardOption = new("ASTPvP_Burst_PlayCardOption");
+
+            internal static void Draw(CustomComboPreset preset)
+            {
+                switch (preset)
+                {
+                    case CustomComboPreset.ASTPvP_Burst_PlayCard:
+                        DrawHorizontalRadioButton(ASTPvP_Burst_PlayCardOption, "Lord and Lady card play",
+                             "Uses Lord and Lady of Crowns when available.", 1);
+
+                        DrawHorizontalRadioButton(ASTPvP_Burst_PlayCardOption, "Lord of Crowns card play",
+                            "Only uses Lord of Crowns when available.", 2);
+
+                        DrawHorizontalRadioButton(ASTPvP_Burst_PlayCardOption, "Lady of Crowns card play",
+                            "Only uses Lady of Crowns when available.", 3);
+
+                        break;
+                }
+            }
 
         internal class ASTPvP_Burst : CustomCombo
         {
