+++ WrathCombo/Combos/PvP/BRDPVP.cs
@@ -1,4 +1,5 @@
 using WrathCombo.CustomComboNS;
+using static WrathCombo.Window.Functions.UserConfig;
 
 namespace WrathCombo.Combos.PvP
 {
@@ -48,7 +49,15 @@ namespace WrathCombo.Combos.PvP
             public const string
                 BRDPvP_HarmonicArrowCharges = "BRDPvP_HarmonicArrowCharges";
 
-        }
+            internal static void Draw(CustomComboPreset preset)
+            {
+                switch (preset)
+                {
+                    case CustomComboPreset.BRDPvP_HarmonicArrow:
+                        DrawSliderInt(1, 4, BRDPvP_HarmonicArrowCharges, "How many Charges to use it at \n 1 charge 8000 damage \n 2 charge 12000 damage \n 3 charge 15000 damage \n 4 charge 17000 damage");
+                        break;
+                }
+            }
 
         internal class BRDPvP_BurstMode : CustomCombo
         {
