+++ WrathCombo/Combos/PvP/NINPVP.cs
@@ -1,4 +1,7 @@
-﻿using WrathCombo.CustomComboNS;
+﻿using ECommons.GameHelpers;
+using WrathCombo.Core;
+using WrathCombo.CustomComboNS;
+using WrathCombo.Window.Functions;
 
 namespace WrathCombo.Combos.PvP
 {
@@ -60,6 +63,48 @@ namespace WrathCombo.Combos.PvP
                 NINPvP_Meisui_AoE = "NINPvP_Meisui_AoE",
                 NINPVP_SeitonTenchu = "NINPVP_SeitonTenchu",
                 NINPVP_SeitonTenchuAoE = "NINPVP_SeitonTenchuAoE";
+
+            internal static void Draw(CustomComboPreset preset)
+            {
+                switch(preset)
+                {
+                    case CustomComboPreset.NINPvP_ST_Meisui:
+                        string descriptionST = "Set the HP percentage to be at or under for the feature to kick in.\n100% is considered to start at 8,000 less than your max HP to prevent wastage.";
+
+                        if (Player.Object != null)
+                        {
+                            uint maxHP = Player.Object.MaxHp <= 8000 ? 0 : Player.Object.MaxHp - 8000;
+                            if (maxHP > 0)
+                            {
+                                int setting = PluginConfiguration.GetCustomIntValue(NINPvP_Meisui_ST);
+                                float hpThreshold = (float)maxHP / 100 * setting;
+
+                                descriptionST += $"\nHP Value to be at or under: {hpThreshold}";
+                            }
+                        }
+
+                        UserConfig.DrawSliderInt(1, 100, NINPvP_Meisui_ST, descriptionST);
+                        break;
+
+                    case CustomComboPreset.NINPvP_AoE_Meisui:
+                        string descriptionAoE = "Set the HP percentage to be at or under for the feature to kick in.\n100% is considered to start at 8,000 less than your max HP to prevent wastage.";
+
+                        if (Player.Object != null)
+                        {
+                            uint maxHP = Player.Object.MaxHp <= 8000 ? 0 : Player.Object.MaxHp - 8000;
+                            if (maxHP > 0)
+                            {
+                                int setting = PluginConfiguration.GetCustomIntValue(NINPvP.Config.NINPvP_Meisui_AoE);
+                                float hpThreshold = (float)maxHP / 100 * setting;
+
+                                descriptionAoE += $"\nHP Value to be at or under: {hpThreshold}";
+                            }
+                        }
+
+                        UserConfig.DrawSliderInt(1, 100, NINPvP.Config.NINPvP_Meisui_AoE, descriptionAoE);
+                        break;
+                }
+            }
         }
 
         internal class NINPvP_ST_BurstMode : CustomCombo
