+++ WrathCombo/Window/Functions/Presets.cs
@@ -314,38 +314,34 @@ namespace WrathCombo.Window.Functions
                         case VPR.JobID: VPR.Config.Draw(preset); break;
                         case WAR.JobID: WAR.Config.Draw(preset); break;
                         case WHM.JobID: WHM.Config.Draw(preset); break;
-                        default: UserConfigItems.Draw(preset, enabled); break;
                     }
                 }
                 else
                 {
                     switch (info.JobID)
                     {
-                        //  case All.JobID: All.Config.Draw(preset); break;
-                        case AST.JobID: AST.Config.Draw(preset); break;
-                        case BLM.JobID: BLM.Config.Draw(preset); break;
-                        case BLU.JobID: BLU.Config.Draw(preset); break;
-                        case BRD.JobID: BRD.Config.Draw(preset); break;
-                        case DNC.JobID: DNC.Config.Draw(preset); break;
-                        //case DOL.JobID: DOL.Config.Draw(preset); break;
-                        case DRG.JobID: DRG.Config.Draw(preset); break;
-                        case DRK.JobID: DRK.Config.Draw(preset); break;
-                        case GNB.JobID: GNB.Config.Draw(preset); break;
-                        case MCH.JobID: MCH.Config.Draw(preset); break;
-                        case MNK.JobID: MNK.Config.Draw(preset); break;
-                        case NIN.JobID: NIN.Config.Draw(preset); break;
-                        case PCT.JobID: PCT.Config.Draw(preset); break;
-                        case PLD.JobID: PLD.Config.Draw(preset); break;
-                        case RPR.JobID: RPR.Config.Draw(preset); break;
+                        case All.JobID: PvPCommon.Config.Draw(preset); break;
+                        case AST.JobID: ASTPvP.Config.Draw(preset); break;
+                        case BLM.JobID: BLMPvP.Config.Draw(preset); break;
+                        case BRD.JobID: BRDPvP.Config.Draw(preset); break;
+                        case DNC.JobID: DNCPvP.Config.Draw(preset); break;
+                        case DRG.JobID: DRGPvP.Config.Draw(preset); break;
+                        case DRK.JobID: DRKPvP.Config.Draw(preset); break;
+                        case GNB.JobID: GNBPvP.Config.Draw(preset); break;
+                        case MCH.JobID: MCHPvP.Config.Draw(preset); break;
+                        case MNK.JobID: MNKPvP.Config.Draw(preset); break;
+                        case NIN.JobID: NINPvP.Config.Draw(preset); break;
+                        case PCT.JobID: PCTPvP.Config.Draw(preset); break;
+                        case PLD.JobID: PLDPvP.Config.Draw(preset); break;
+                        case RPR.JobID: RPRPvP.Config.Draw(preset); break;
                         case RDM.JobID: RDMPvP.Config.Draw(preset); break;
-                        case SAM.JobID: SAM.Config.Draw(preset); break;
-                        case SCH.JobID: SCH.Config.Draw(preset); break;
-                        case SGE.JobID: SGE.Config.Draw(preset); break;
-                        case SMN.JobID: SMN.Config.Draw(preset); break;
-                        case VPR.JobID: VPR.Config.Draw(preset); break;
-                        case WAR.JobID: WAR.Config.Draw(preset); break;
-                        case WHM.JobID: WHM.Config.Draw(preset); break;
-                        default: UserConfigItems.Draw(preset, enabled); break;
+                        case SAM.JobID: SAMPvP.Config.Draw(preset); break;
+                        case SCH.JobID: SCHPvP.Config.Draw(preset); break;
+                        case SGE.JobID: SGEPvP.Config.Draw(preset); break;
+                        case SMN.JobID: SMNPvP.Config.Draw(preset); break;
+                        case VPR.JobID: VPRPvP.Config.Draw(preset); break;
+                        case WAR.JobID: WARPvP.Config.Draw(preset); break;
+                        case WHM.JobID: WHMPvP.Config.Draw(preset); break;
                     }
                 }
